requirements.txt
attrs==22.1.0
Automat==20.2.0
boto3==1.24.49
botocore==1.27.49
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
docopt==0.6.2
feedparser==6.0.10
filelock==3.7.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
parsel==1.6.0
pipreqs==0.4.11
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pymongo==4.2.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
s3transfer==0.6.0
Scrapy==2.6.2
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
text-unidecode==1.3
tldextract==3.3.1
Twisted==22.4.0
typing_extensions==4.3.0
unicode-slugify==0.1.5
Unidecode==1.3.4
urllib3==1.26.11
w3lib==1.22.0
yarg==0.1.9
zope.interface==5.4.0
